requirements.txt
# Web framework and server
fastapi
uvicorn
python-multipart

# Frontend integration and HTTP
streamlit
requests
beautifulsoup4

# Finance data
yfinance

# LLMs and RAG stack
openai
langchain
langchain-core>=0.3.0
langchain-community>=0.0.20
langchain-openai>=0.0.10
sentence-transformers
faiss-cpu

# Whisper for speech-to-text
git+https://github.com/openai/whisper.git

# TTS and audio
pyttsx3
gTTS
# Env variables
python-dotenv

# Together API
together
